// apps/web/src/shared/api/generated/model/statisticsControllerGetOverviewParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * book-nest API
 * REST API for the book-nest project
 * OpenAPI spec version: 1.0
 */
import type { StatisticsControllerGetOverviewCompare } from "./statisticsControllerGetOverviewCompare";
import type { StatisticsControllerGetOverviewPeriod } from "./statisticsControllerGetOverviewPeriod";

export type StatisticsControllerGetOverviewParams = {
  compare?: StatisticsControllerGetOverviewCompare;
  /**
   * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
   */
  from?: string;
  period?: StatisticsControllerGetOverviewPeriod;
  /**
   * @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
   */
  to?: string;
  /**
   * @minimum 1000
   * @maximum 9999
   */
  year?: number;
};